Move typeInfo to package
