+++ internal/src/main/java/online/sharedtype/processor/domain/ClassDef.java
@@ -3,6 +3,9 @@ package online.sharedtype.processor.domain;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.SuperBuilder;
+import online.sharedtype.processor.domain.type.ConcreteTypeInfo;
+import online.sharedtype.processor.domain.type.TypeInfo;
+import online.sharedtype.processor.domain.type.TypeVariableInfo;
 
 import java.util.ArrayList;
 import java.util.HashMap;
