+++ internal/src/main/java/online/sharedtype/processor/domain/FieldComponentInfo.java
@@ -2,6 +2,7 @@ package online.sharedtype.processor.domain;
 
 import lombok.Builder;
 import lombok.Setter;
+import online.sharedtype.processor.domain.type.TypeInfo;
 
 import javax.lang.model.element.Modifier;
 import java.util.Set;
