+++ it/java17/src/test/java/online/sharedtype/it/MapClassIntegrationTest.java
@@ -1,7 +1,7 @@
 package online.sharedtype.it;
 
 import online.sharedtype.processor.domain.ClassDef;
-import online.sharedtype.processor.domain.ConcreteTypeInfo;
+import online.sharedtype.processor.domain.type.ConcreteTypeInfo;
 import online.sharedtype.processor.domain.Constants;
 import org.junit.jupiter.api.Test;
