+++ it/java17/src/test/java/online/sharedtype/it/OptionalTypeIntegrationTest.java
@@ -1,8 +1,8 @@
 package online.sharedtype.it;
 
-import online.sharedtype.processor.domain.ArrayTypeInfo;
+import online.sharedtype.processor.domain.type.ArrayTypeInfo;
 import online.sharedtype.processor.domain.ClassDef;
-import online.sharedtype.processor.domain.ConcreteTypeInfo;
+import online.sharedtype.processor.domain.type.ConcreteTypeInfo;
 import online.sharedtype.processor.domain.Constants;
 import org.junit.jupiter.api.Test;
