+++ it/java8/src/test/java/online/sharedtype/it/JavaClassIntegrationTest.java
@@ -1,9 +1,9 @@
 package online.sharedtype.it;
 
 import online.sharedtype.processor.domain.ClassDef;
-import online.sharedtype.processor.domain.ConcreteTypeInfo;
+import online.sharedtype.processor.domain.type.ConcreteTypeInfo;
 import online.sharedtype.processor.domain.FieldComponentInfo;
-import online.sharedtype.processor.domain.TypeVariableInfo;
+import online.sharedtype.processor.domain.type.TypeVariableInfo;
 import online.sharedtype.it.support.TypeDefDeserializer;
 import org.junit.jupiter.api.Test;
