+++ processor/src/main/java/online/sharedtype/processor/resolver/SubtypeResolver.java
@@ -1,9 +1,9 @@
 package online.sharedtype.processor.resolver;
 
 import online.sharedtype.processor.domain.ClassDef;
-import online.sharedtype.processor.domain.ConcreteTypeInfo;
+import online.sharedtype.processor.domain.type.ConcreteTypeInfo;
 import online.sharedtype.processor.domain.TypeDef;
-import online.sharedtype.processor.domain.TypeInfo;
+import online.sharedtype.processor.domain.type.TypeInfo;
 
 import java.util.List;
