+++ processor/src/main/java/online/sharedtype/processor/resolver/TypeResolver.java
@@ -2,7 +2,7 @@ package online.sharedtype.processor.resolver;
 
 import online.sharedtype.SharedType;
 import online.sharedtype.processor.domain.TypeDef;
-import online.sharedtype.processor.domain.TypeInfo;
+import online.sharedtype.processor.domain.type.TypeInfo;
 import online.sharedtype.processor.context.Context;
 import online.sharedtype.processor.parser.TypeDefParser;
