+++ processor/src/main/java/online/sharedtype/processor/writer/converter/type/AbstractTypeExpressionConverter.java
@@ -3,14 +3,14 @@ package online.sharedtype.processor.writer.converter.type;
 import lombok.RequiredArgsConstructor;
 import online.sharedtype.processor.context.Config;
 import online.sharedtype.processor.context.Context;
-import online.sharedtype.processor.domain.ArrayTypeInfo;
+import online.sharedtype.processor.domain.type.ArrayTypeInfo;
 import online.sharedtype.processor.domain.ClassDef;
-import online.sharedtype.processor.domain.ConcreteTypeInfo;
+import online.sharedtype.processor.domain.type.ConcreteTypeInfo;
 import online.sharedtype.processor.domain.Constants;
-import online.sharedtype.processor.domain.DateTimeInfo;
+import online.sharedtype.processor.domain.type.DateTimeInfo;
 import online.sharedtype.processor.domain.TypeDef;
-import online.sharedtype.processor.domain.TypeInfo;
-import online.sharedtype.processor.domain.TypeVariableInfo;
+import online.sharedtype.processor.domain.type.TypeInfo;
+import online.sharedtype.processor.domain.type.TypeVariableInfo;
 import online.sharedtype.processor.support.annotation.SideEffect;
 import online.sharedtype.processor.support.exception.SharedTypeInternalError;
