+++ processor/src/test/java/online/sharedtype/processor/domain/ClassDefTest.java
@@ -1,5 +1,8 @@
 package online.sharedtype.processor.domain;
 
+import online.sharedtype.processor.domain.type.ArrayTypeInfo;
+import online.sharedtype.processor.domain.type.ConcreteTypeInfo;
+import online.sharedtype.processor.domain.type.TypeVariableInfo;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
