+++ processor/src/test/java/online/sharedtype/processor/resolver/OptionalTypeResolverTest.java
@@ -1,9 +1,9 @@
 package online.sharedtype.processor.resolver;
 
 import online.sharedtype.processor.context.ContextMocks;
-import online.sharedtype.processor.domain.ArrayTypeInfo;
+import online.sharedtype.processor.domain.type.ArrayTypeInfo;
 import online.sharedtype.processor.domain.ClassDef;
-import online.sharedtype.processor.domain.ConcreteTypeInfo;
+import online.sharedtype.processor.domain.type.ConcreteTypeInfo;
 import online.sharedtype.processor.domain.Constants;
 import online.sharedtype.processor.domain.FieldComponentInfo;
 import org.junit.jupiter.api.BeforeEach;
