+++ processor/src/test/java/online/sharedtype/processor/resolver/ReferenceResolverTest.java
@@ -2,8 +2,8 @@ package online.sharedtype.processor.resolver;
 
 
 import online.sharedtype.processor.domain.ClassDef;
-import online.sharedtype.processor.domain.ConcreteTypeInfo;
-import online.sharedtype.processor.domain.TypeVariableInfo;
+import online.sharedtype.processor.domain.type.ConcreteTypeInfo;
+import online.sharedtype.processor.domain.type.TypeVariableInfo;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
