+++ processor/src/test/java/online/sharedtype/processor/resolver/SubtypeResolverTest.java
@@ -1,7 +1,7 @@
 package online.sharedtype.processor.resolver;
 
 import online.sharedtype.processor.domain.ClassDef;
-import online.sharedtype.processor.domain.ConcreteTypeInfo;
+import online.sharedtype.processor.domain.type.ConcreteTypeInfo;
 import online.sharedtype.processor.domain.TypeDef;
 import org.junit.jupiter.api.Test;
