+++ processor/src/test/java/online/sharedtype/processor/writer/converter/ConversionUtilsTest.java
@@ -1,7 +1,7 @@
 package online.sharedtype.processor.writer.converter;
 
 import online.sharedtype.processor.domain.ClassDef;
-import online.sharedtype.processor.domain.ConcreteTypeInfo;
+import online.sharedtype.processor.domain.type.ConcreteTypeInfo;
 import online.sharedtype.processor.domain.FieldComponentInfo;
 import org.junit.jupiter.api.Test;
